% Generated by roxygen2 (4.0.2): do not edit by hand
\name{collapseDuplicatesWithinString}
\alias{collapseDuplicatesWithinString}
\title{Collapse duplicate enteries within a string}
\usage{
collapseDuplicatesWithinString(str, sep = "; ")
}
\arguments{
\item{str}{Character vector in which individual strings are separated by
`sep`, which by default is "; "}

\item{sep}{separator of enteries within individual strings}
}
\value{
A character vector with removed duplicated enteries.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

