% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAllBanks}
\alias{getAllBanks}
\title{Convinience function: Get post-processed list of all banks involved
in any of EBA or ECB tests.}
\usage{
getAllBanks(outfile = NULL)
}
\arguments{
\item{outfile}{path of the output csv file. By default, no file is written.}
}
\value{
data.table with banks (containing Bankscope lookup information).
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

