% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadManuallyLinkedBanks}
\alias{loadManuallyLinkedBanks}
\title{Load a table of manually linked banks that participated in EBA and
ECB tests}
\usage{
loadManuallyLinkedBanks()
}
\value{
data.table with banks that participated in EBA and ECB tests.
}
\description{
This function loads a table of banks that participated in EBA or ECB
tests. Due to a lack of consistency in naming and identification of banks
across different test, the participating banks needed to be manually linked
acroos the tests. I also manually link the banks with their identifying
information in Bankscope.
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

