% Generated by roxygen2 (4.0.2): do not edit by hand
\name{selectBanksInTest}
\alias{selectBanksInTest}
\title{Get a selection of banks involved in a specified test}
\usage{
selectBanksInTest(testName = "AQR2014", outfile = NULL)
}
\arguments{
\item{testName}{Name of the test. Currently supported: "AQR2014",
"EBARECAP2012", "EBATRANSPARENCY2013", "STRESSTEST2011", "STRESSTEST2014"}

\item{outfile}{path of the output csv file. By default, no file is written.}
}
\value{
data.table with banks involved in the querried test.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

