% Generated by roxygen2 (4.0.2): do not edit by hand
\name{selectUniqueBSIndex}
\alias{selectUniqueBSIndex}
\title{Bankscope Index Selector}
\usage{
selectUniqueBSIndex(bsIndexCol)
}
\arguments{
\item{bsIndexCol}{Character vector in which each entry has the following
strucuture: [BS_INDEX_1] ([CONSOL_1]); [BS_INDEX_2] ([CONSOL_2]);...}
}
\value{
Numeric vector of selected Bankscope Indices
}
\description{
A given bank typically has multiple (manually) matched Bankscope
identifiers, corresponding to different levels of accounting
consolidation. For a given bank, this function selects an index at the
higest available level of consolidation.
}
\details{
TODO: Describe Bankscope consolidation concepts
}
\author{
Janko Cizel
}

